Banish ExternalProject; allowing build with external GraftNetwork
ExternalProject is a dirty hack for include a cmake project, and has
many drawbacks (among other things it means cmake can't properly track
when things need to be rebuilt because of changes in dependent
projects).

It's also completely unnecessary (with a couple of fixes to
GraftNetwork).

This also adds a EXTERNAL_CRYPTONODE option so that the build can be
done against an external copy of GraftNetwork installed elsewhere rather
than the embedded one (defaults to OFF, of course).  Actually doing this
also requires adding appropriate -I and -L options with something like
`-DCMAKE_CXX_FLAGS='-I/some/include -I/some/lib` when invoking cmake.